ByteBuffer bounds checking fixes
This fixes a number of bounds checking issues, including changing the behavior of out of bounds access. Now if get is passed a negative value or a value beyond the size of the buffer, it explodes. If set is passed a negative value it also explodes, but values passed the end of the buffer continue to implicitly resize.