Ensure we always mark the flush checkpoint, even if we are waiting for writable() to be called.
Previously we did not correctly mark the flush checkpoint if a call to writable() by the EventLoop was pending. This could have the affect that when the user calls flush while we already registered for InterestedEvent.write or InterestedEvent.all not all messages in the PendingWrite datastructure were correctly marked as flushed and so not written at the end when the writable call happens.