Remove use of DispatchQueue which was used to workaround https://bugs.swift.org/browse/SR-5145.
The use of DispatchQueue introduced a lot of overhead because it dispatched th ework to many threads. This could lead to have only one thread doing the IO while many others the actual deinit.

As we already limit the number of PendingWrite instances that will be written via writev to writevLimit we only need to take special action on failAll(...) which may possibly release thousands of PendingWrite instances and so could trigger a stackoverflow.