Ensure we not segfault when http pipelining is used with the HTTP1Server
Motivation:

We need to ensure we not use a shared count in the HTTPHandler as we may receive requested faster then we are able to flush out the responses and both act on the count.

Modifications:

Define continuousCount in the function body itself.

Result:

No more segfault possible even if pipelining was used.