Replace precondition with assert when checks for an implementation bug.
Motivation:

We used precondition for a few cases where it really checks for our own implementation. We should use assert here as there is nothing the user really can do to trigger it.

Modifications:

Replace precondition with assert.

Result:

Less overhead when in release mode and more correct in general.