Rename PriorityQueue's T to Element and iterator to Iterator, avoiding need for typealiases. (#91)
Motivation:

We've found following these conventions in the standard library simplifies
the code significantly if you create a lot of generic types that conform
to the std lib protocols.

Modifications:

Change the placeholder T to Element and remove now-unnecessary typealias.
Switch the name of the iterator to just Iterator, and have it re-use
the outer type's Element placeholder. Remove now-unnecessary typealiases.
Add deprecated typealiases to make this a non-breaking change.

Result:

Should be a non-functional change. Explicit typing outside the PriorityQueue
type would need to change, but there wasn't any.