Fix test compilation on Linux issue #110
Motivation:

It would be great if the test would compile and run
on Linux, so that it can pass CI.

Modifications:

Do not fill sin6_len on Linux.
And fix the other bug in the test, which was filling
the sin6_len on Darwin w/ the wrong value 🤫

Result:

Compile on Linux and then hopefully passes the CI,
finally.

In Docker:
test_1               | Test Case 'SocketAddressTest.testIn6AddrDescriptionWorks' started at 2018-03-08 09:37:54.447
test_1               | Test Case 'SocketAddressTest.testIn6AddrDescriptionWorks' passed (0.0 seconds)