Add EventLoopPromise.eventLoop and add docs
Motivation:

It was not possible to access the EventLoop directly that was used to create the EventLoopPromise which had the effect that the user needed to first get the futureResult and then access the EventLoop. Also there were no docs.

Modifications:

- Add property to access the EventLoop directly.
- Add missing docs.

Result:

Easier way to access EventLoop and also docs.