Add support for getting a handler of a given type.
Motivation:

It is often useful to find a handler of a given type even when you
aren't holding a reference to that handler. This patch allows us to
do that.

Modifications:

Add context(handlerType:) function to the channel pipeline.

Result:

Users can find handlers of a given type in a pipeline.