Tail and Head handler of ChannelPipeline should always remain in the ChannelPipeline
Motivation:

We must ensure we always keep the Tail and Head handler in the ChannelPipeline, even once we close the Channel (and so call removeHandlers() ). This is because Channel.ChannelCore.channelRead0(...) (which is called from the HeadChannelHandler) may do special handling to release messages for example. This is the case for our ServerSocketChannel which received SocketChannel instances and so register these. The problem is if we don't do this the user will never see the SocketChannel and so has no chance to close it. This will result in a leak of the underlying file descriptor and so crash.

Modifications:

- Ensure we never remove the tail and head handlers from the ChannelPipeline.
- Also skip tail and head when searching for a specific handler as these are internal.
- skip tail and head handler when create the debug string for the ChannelPipeline as again these are internal and not useful for the user

Result:

No more leak when the ServerSocketChannel is closed while still some SocketChannel are flowing through the ChannelPipeline by channelRead.