Ensure we always call Channel.close() if a channel is fired through the ChannelPipeline and the pipeline is already destroyed.
Motivation:

If a Channel is closed we also destroy the ChannelPipeline. For this we remove all handlers from it. The problem is that once this is done there is no more posibility for the user to intercept the fireChannelRead(...) and so Channels may leak if these are passed into it (as done by ServerSocketChannel).

Modifications:

Close the Channel when the pipeline was destroyed already.

Result:

No more leaks of Channels accepted by ServerSocketChannel.