add a simple leak detector (#255)
Motivation:

Memory leaks are rare in Swift but they can happen. If we run a
realistic workload and the number of `malloc`s and `free`s is very close
to 0, we can be fairly sure we didn't have any (too bad) leaks.
Therefore this adds a simple leaks detector.

Modifications:

Add a leak detector.

Result:

We should now detect leaks even if the total number of allocations
doesn't increase.