make allocation counter test fail on error in subshell
Motivation:

In shell, subshells don't inherit the shell flags (such as `set -eu`)
and therefore need to set them again. The allocation counter test uses a
subshell but forgot to `set -eu` in the subshell.

Modifications:

`set -eu` in the subshell for the allocation counter tests.

Result:

if there's an error building the test program the test now fails.