Rework NIOWebSocketServer main to bind host similar to other servers
Motivation:

To move NIOWebSocketServer inline with how other servers bind to host.

Modifications:

The main file in NIOWebSocketServer now accepts command line arguments
and binds to a passed in host and port like the other server examples do.

Result:

The result of my change will bring the NIOWebSocketServer more inline
with how the NIOHTTP1 and NIOChat servers examples are setup.