Don't accidentally fail leak detection tests.
Motivation:

The channel leak detection test currently gates its progress based on
the firing of the promise passed to close(). That promise fires way, way
before the channel stops scheduling work, which means that it's highly
likely the event loop will still be holding a ref to the channel shortly
after that promise fires.

That causes the test to fail. Failing tests are bad.

Modifications:

Wait for the closeFuture instead. This fires at the end of the last
callback a socket channel ever schedules, so will fire just before the
channel drops out of scope.

Result:

Less flaky tests.