Don't race waiting for a nil reference. (#289)
Motivation:

Unfortunately we cannot guarantee that a weak ref is going to be nil in
a timely fashion if a strong reference to the object exists in another
thread. This makes us fail tests on slow machines, which is bad.

Modifications:

Wait for up to a second for the other thread to get its house in order
before we give up.

Result:

Hopefully I'll never have to "fix" this test again.