add 'info' mode for integration tests
Motivation:

Sometimes it's useful if the integration tests can output something that
isn't always useful. `-v` for verbose is way too verbose so this adds a
`-i` flag to the integration tests that lets 'info' messages through.

Modifications:

- add `-i` flags to integration tests
- make use of it in the integration counter tests

Result:

if we enable `-i` in CI, then we can constantly monitor the number of
allocations we do even if we don't fail the tests.