Make IdleStateHandler.IdleStateEvent public (#319)
IdleStateHandler.IdleStateEvent is currently internal. Make it public.

Motivation:

The IdleStateEvent is delivered to the user viapublic func userInboundEventTriggered(ctx: ChannelHandlerContext, event: Any). In this method, it must be possible to have a check on the type of event. IdleStateEvent must be public.

Modifications:

Change access level of IdleStateHandler.IdleStateEvent to public.

Result:

IdleStateEvent will become a part of the NIO API.