Hard-fail on invalid write types in core channels.
Motivation:

There is relatively little benefit in allowing the user to send invalid
data types on a given channel. Currently we tolerate this by simply
failing the promise, but this is not really ideal: if the user never sets
a promise here they never even find out that we bailed on sending their
data. Given that we attempt to be type safe everywhere else, we should
just blow up immediately if the user attempts to write an invalid data
type.

Modifications:

Removed the nice gentle type casts, replaced them with explody ones.

Result:

Faster failure in error cases.